:truck: Refs #612 - Move component selection and component context
