--- src/openforms/js/components/admin/form_design/logic/Context.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-
-const ComponentsContext = React.createContext([]);
-ComponentsContext.displayName = 'ComponentsContext';
-
-export { ComponentsContext };
