+++ src/openforms/js/components/admin/forms/Context.js
@@ -1,7 +1,9 @@
 import React from 'react';
 
 const PrefixContext = React.createContext('');
+const ComponentsContext = React.createContext([]);
 
 PrefixContext.displayName = 'PrefixContext';
+ComponentsContext.displayName = 'ComponentsContext';
 
-export { PrefixContext };
+export { PrefixContext, ComponentsContext };
